+++ media/smart-card/src/extractors/common/__tests__/utils.test.ts
@@ -1,11 +1,6 @@
-import {
-  extractUrlFromIconJsonLd,
-  extractUrlFromLinkJsonLd,
-  extractPersonFromJsonLd,
-} from '../utils';
+import { extractUrlFromIconJsonLd, extractPersonFromJsonLd } from '../utils';
 import {
   TEST_STRING,
-  TEST_ARRAY,
   TEST_URL,
   TEST_LINK,
   TEST_IMAGE,
@@ -16,24 +11,6 @@ import {
 
 describe('extractor.utils', () => {
   afterEach(() => jest.clearAllMocks());
-
-  describe('extractUrlFromLinkJsonLd', () => {
-    it('returns raw string', () => {
-      expect(extractUrlFromLinkJsonLd(TEST_STRING)).toBe(TEST_STRING);
-    });
-
-    it('returns undefined for empty array', () => {
-      expect(extractUrlFromLinkJsonLd([])).toBe(undefined);
-    });
-
-    it('returns href of first element of array', () => {
-      expect(extractUrlFromLinkJsonLd(TEST_ARRAY)).toBe(TEST_URL);
-    });
-
-    it('returns href of passed element', () => {
-      expect(extractUrlFromLinkJsonLd(TEST_LINK)).toBe(TEST_URL);
-    });
-  });
 
   describe('extractUrlFromIconJsonLd', () => {
     it('returns raw string', () => {