+++ media/smart-card/src/extractors/common/preview/index.ts
@@ -1,3 +1,3 @@
+export { extractPreview } from '@atlaskit/linking-common';
+export type { LinkPreview } from '@atlaskit/linking-common';
 export { extractImage } from './extractImage';
-export { extractPreview } from './extractPreview';
-export type { LinkPreview } from './extractPreview';