+++ media/smart-card/src/providers/editor/types.ts
@@ -17,24 +17,3 @@ type Provider = {
 export type ORSProvidersResponse = {
   providers: Provider[];
 };
-
-export interface InlineCardAdf {
-  type: 'inlineCard';
-  attrs: {
-    url: string;
-  };
-}
-export interface BlockCardAdf {
-  type: 'blockCard';
-  attrs: {
-    url: string;
-  };
-}
-export interface EmbedCardAdf {
-  type: 'embedCard';
-  attrs: {
-    url: string;
-    layout: 'wide';
-  };
-}
-export type CardAdf = InlineCardAdf | BlockCardAdf | EmbedCardAdf;